Finished Project (made sure price was numeric)

+++ db/migrate/20220203220120_add_price_to_book.rb
@@ -1,5 +1,5 @@
 class AddPriceToBook < ActiveRecord::Migration[6.1]
   def change
-    add_column :books, :price, :float
+    add_column :books, :price, :numeric
   end
 end
